Remove old Variable, move new_variable to variable
