+++ src/easyscience/Constraints.py
@@ -195,9 +195,9 @@ class NumericConstraint(ConstraintBase):
 
     def _parse_operator(self, obj: V, *args, **kwargs) -> Number:
         ## TODO clean when full move to new_variable
-        import easyscience.Objects.new_variable.parameter
+        import easyscience.Objects.variable.parameter
 
-        if isinstance(obj, easyscience.Objects.new_variable.parameter.Parameter):
+        if isinstance(obj, easyscience.Objects.variable.parameter.Parameter):
             value = obj.value_no_call_back
         else:
             value = obj.raw_value
@@ -259,9 +259,9 @@ class SelfConstraint(ConstraintBase):
 
     def _parse_operator(self, obj: V, *args, **kwargs) -> Number:
         ## TODO clean when full move to new_variable
-        import easyscience.Objects.new_variable.parameter
+        import easyscience.Objects.variable.parameter
 
-        if isinstance(obj, easyscience.Objects.new_variable.parameter.Parameter):
+        if isinstance(obj, easyscience.Objects.variable.parameter.Parameter):
             value = obj.value_no_call_back
         else:
             value = obj.raw_value
@@ -323,9 +323,9 @@ class ObjConstraint(ConstraintBase):
 
     def _parse_operator(self, obj: V, *args, **kwargs) -> Number:
         ## TODO clean when full move to new_variable
-        import easyscience.Objects.new_variable.parameter
+        import easyscience.Objects.variable.parameter
 
-        if isinstance(obj, easyscience.Objects.new_variable.parameter.Parameter):
+        if isinstance(obj, easyscience.Objects.variable.parameter.Parameter):
             value = obj.value_no_call_back
         else:
             value = obj.raw_value
@@ -417,13 +417,13 @@ class MultiObjConstraint(ConstraintBase):
         self.external = True
 
     def _parse_operator(self, independent_objs: List[V], *args, **kwargs) -> Number:
-        import easyscience.Objects.new_variable.parameter
+        import easyscience.Objects.variable.parameter
 
         in_str = ''
         value = None
         for idx, obj in enumerate(independent_objs):
             ## TODO clean when full move to new_variable
-            if isinstance(obj, easyscience.Objects.new_variable.parameter.Parameter):
+            if isinstance(obj, easyscience.Objects.variable.parameter.Parameter):
                 self.aeval.symtable['p' + str(self.independent_obj_ids[idx])] = obj.value_no_call_back
             else:
                 self.aeval.symtable['p' + str(self.independent_obj_ids[idx])] = obj.raw_value
@@ -485,14 +485,14 @@ class FunctionalConstraint(ConstraintBase):
             self.external = True
 
     def _parse_operator(self, obj: V, *args, **kwargs) -> Number:
-        import easyscience.Objects.new_variable.parameter
+        import easyscience.Objects.variable.parameter
 
         self.aeval.symtable[f'f{id(self.function)}'] = self.function
         value_str = f'r_value = f{id(self.function)}('
         if isinstance(obj, list):
             for o in obj:
                 ## TODO clean when full move to new_variable
-                if isinstance(o, easyscience.Objects.new_variable.parameter.Parameter):
+                if isinstance(o, easyscience.Objects.variable.parameter.Parameter):
                     value_str += f'{o.value_no_call_back},'
                 else:
                     value_str += f'{o.raw_value},'
@@ -500,7 +500,7 @@ class FunctionalConstraint(ConstraintBase):
             value_str = value_str[:-1]
         else:
             ## TODO clean when full move to new_variable
-            if isinstance(obj, easyscience.Objects.new_variable.parameter.Parameter):
+            if isinstance(obj, easyscience.Objects.variable.parameter.Parameter):
                 value_str += f'{obj.value_no_call_back}'
             else:
                 value_str += f'{obj.raw_value}'
