+++ src/easyscience/Objects/old_Variable.py
