+++ src/easyscience/Objects/variable/__init__.py
