+++ src/easyscience/Objects/variable/descriptor_base.py
