+++ src/easyscience/Objects/variable/descriptor_bool.py
