+++ src/easyscience/Objects/variable/descriptor_number.py
