+++ src/easyscience/Objects/variable/descriptor_str.py
