+++ src/easyscience/Objects/variable/parameter.py
