+++ src/easyscience/fitting/minimizers/minimizer_base.py
@@ -20,7 +20,7 @@ from easyscience.Constraints import ObjConstraint
 
 # causes circular import when Parameter is imported
 # from easyscience.Objects.ObjectClasses import BaseObj
-from easyscience.Objects.new_variable import Parameter
+from easyscience.Objects.variable import Parameter
 
 from ..available_minimizers import AvailableMinimizers
 from .utils import FitError
