+++ src/easyscience/fitting/minimizers/minimizer_lmfit.py
@@ -15,7 +15,7 @@ from lmfit.model import ModelResult
 
 # causes circular import when Parameter is imported
 # from easyscience.Objects.ObjectClasses import BaseObj
-from easyscience.Objects.new_variable import Parameter
+from easyscience.Objects.variable import Parameter
 
 from ..available_minimizers import AvailableMinimizers
 from .minimizer_base import MINIMIZER_PARAMETER_PREFIX
