+++ tests/unit_tests/Objects/test_BaseObj.py
@@ -19,8 +19,8 @@ import pytest
 
 import easyscience
 from easyscience.Objects.ObjectClasses import BaseObj
-from easyscience.Objects.new_variable import DescriptorNumber
-from easyscience.Objects.new_variable import Parameter
+from easyscience.Objects.variable import DescriptorNumber
+from easyscience.Objects.variable import Parameter
 from easyscience.Utils.io.dict import DictSerializer
 from easyscience import global_object
 
@@ -357,7 +357,7 @@ def test_subclassing():
     from typing import ClassVar
 
     from easyscience.models.polynomial import Line
-    from easyscience.Objects.new_variable import Parameter
+    from easyscience.Objects.variable import Parameter
 
     class L2(Line):
         diff: ClassVar[Parameter]
