+++ tests/unit_tests/global_object/test_global_object.py
@@ -1,6 +1,6 @@
 import easyscience
 from easyscience.global_object.global_object import GlobalObject
-from easyscience.Objects.new_variable.descriptor_bool import DescriptorBool
+from easyscience.Objects.variable.descriptor_bool import DescriptorBool
 
 class TestGlobalObject:
     def test_init(self):
