+++ tests/unit_tests/utils/io_tests/test_json.py
@@ -9,7 +9,7 @@ import pytest
 
 from easyscience.Utils.io.json import JsonDataSerializer
 from easyscience.Utils.io.json import JsonSerializer
-from easyscience.Objects.new_variable import DescriptorNumber
+from easyscience.Objects.variable import DescriptorNumber
 
 from .test_core import A
 from .test_core import check_dict
